feat: make main playlist filterable by album & artist
Move the function from the media page to the playlist service
